+++ NEWS.md
@@ -1,3 +1,5 @@
+# epidict 0.1.0
+
 # epidict 0.0.0.9001
 
 * Updating to have all surveys in ODK format 
@@ -12,9 +14,6 @@
 * Add vignettes and README describing the dictionaries
 * Release to CRAN
 * Fix bug in generating elegible and interviewed for Mortality and Nurition surveys
-
-# epidict 0.0.0.9000
-
 * Added a `NEWS.md` file to track changes to the package.
 * Generalize `gen_data()` function to be able to take different organization
   dictionaries in the future.